requirements.txt
# Safety_Flash Pro - Mining Incident Report System
# Core dependencies

# Web Framework
streamlit>=1.32.0

# Database
supabase>=2.3.0

# AI/ML (Solo Gemini)
google-genai>=1.0.0

# PDF Generation
reportlab>=4.1.0
Pillow>=10.2.0

# Audio Processing
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Email
# (uses built-in smtplib)

# Date/Time
python-dateutil>=2.8.2
